+++ scout/constants/__init__.py
@@ -108,17 +108,13 @@ PAR_COORDINATES = {
         "X": IntervalTree(
             [Interval(60001, 2699521, "par1"), Interval(154931044, 155260561, "par2")]
         ),
-        "Y": IntervalTree(
-            [Interval(10001, 2649521, "par1"), Interval(59034050, 59363567, "par2")]
-        ),
+        "Y": IntervalTree([Interval(10001, 2649521, "par1"), Interval(59034050, 59363567, "par2")]),
     },
     "38": {
         "X": IntervalTree(
             [Interval(10001, 2781480, "par1"), Interval(155701383, 156030896, "par2")]
         ),
-        "Y": IntervalTree(
-            [Interval(10001, 2781480, "par1"), Interval(56887903, 57217416, "par2")]
-        ),
+        "Y": IntervalTree([Interval(10001, 2781480, "par1"), Interval(56887903, 57217416, "par2")]),
     },
 }