+++ scout/parse/case.py
@@ -105,9 +105,7 @@ def parse_case_data(
 
     if "gene_panels" in config_data:
         # handle whitespace in gene panel names
-        config_data["gene_panels"] = [
-            panel.strip() for panel in config_data["gene_panels"]
-        ]
+        config_data["gene_panels"] = [panel.strip() for panel in config_data["gene_panels"]]
         config_data["default_gene_panels"] = [
             panel.strip() for panel in config_data["default_gene_panels"]
         ]
@@ -133,9 +131,7 @@ def parse_case_data(
         LOG.info("Adding SMN info from {}.".format(config_data["smn_tsv"]))
         add_smn_info(config_data)
 
-    config_data["vcf_cancer"] = (
-        vcf_cancer if vcf_cancer else config_data.get("vcf_cancer")
-    )
+    config_data["vcf_cancer"] = vcf_cancer if vcf_cancer else config_data.get("vcf_cancer")
     config_data["vcf_cancer_sv"] = (
         vcf_cancer_sv if vcf_cancer_sv else config_data.get("vcf_cancer_sv")
     )
@@ -147,9 +143,7 @@ def parse_case_data(
     config_data["rank_model_version"] = str(config_data.get("rank_model_version", ""))
     config_data["rank_score_threshold"] = config_data.get("rank_score_threshold", 0)
 
-    config_data["sv_rank_model_version"] = str(
-        config_data.get("sv_rank_model_version", "")
-    )
+    config_data["sv_rank_model_version"] = str(config_data.get("sv_rank_model_version", ""))
 
     config_data["track"] = config_data.get("track", "rare")
     if config_data["vcf_cancer"] or config_data["vcf_cancer_sv"]:
@@ -187,9 +181,7 @@ def add_smn_info(config_data):
             ]:
                 ind[key] = smn_info[ind_id][key]
         except KeyError as e:
-            LOG.warning(
-                "Individual {} has no SMN info to update: {}.".format(ind_id, e)
-            )
+            LOG.warning("Individual {} has no SMN info to update: {}.".format(ind_id, e))
 
 
 def add_smn_info_case(case_data):
@@ -221,9 +213,7 @@ def add_smn_info_case(case_data):
             ]:
                 ind[key] = smn_info[ind_id][key]
         except KeyError as e:
-            LOG.warning(
-                "Individual {} has no SMN info to update: {}.".format(ind_id, e)
-            )
+            LOG.warning("Individual {} has no SMN info to update: {}.".format(ind_id, e))
 
 
 def add_peddy_information(config_data):
@@ -264,9 +254,7 @@ def add_peddy_information(config_data):
         ind = analysis_inds[ind_id]
         # Check if peddy has inferred the ancestry
         if ind_id in ped_info:
-            ind["predicted_ancestry"] = ped_info[ind_id].get(
-                "ancestry-prediction", "UNKNOWN"
-            )
+            ind["predicted_ancestry"] = ped_info[ind_id].get("ancestry-prediction", "UNKNOWN")
         # Check if peddy has inferred the sex
         if ind_id in sex_check:
             if sex_check[ind_id]["error"]:
@@ -389,9 +377,7 @@ def parse_individual(sample):
     ind_info["tiddit_coverage_wig"] = sample.get(
         "tiddit_coverage_wig", sample.get("tiddit_coverage_wig")
     )
-    ind_info["upd_regions_bed"] = sample.get(
-        "upd_regions_bed", sample.get("upd_regions_bed")
-    )
+    ind_info["upd_regions_bed"] = sample.get("upd_regions_bed", sample.get("upd_regions_bed"))
     ind_info["upd_sites_bed"] = sample.get("upd_sites_bed", sample.get("upd_sites_bed"))
     ind_info["mt_bam"] = sample.get("mt_bam")
     ind_info["analysis_type"] = sample.get("analysis_type")
@@ -407,9 +393,7 @@ def parse_individual(sample):
     ind_info["smn2delta78_cn"] = sample.get("smn2delta78_cn", None)
     ind_info["smn_27134_cn"] = sample.get("smn_27134_cn", None)
 
-    ind_info["capture_kits"] = (
-        [sample.get("capture_kit")] if "capture_kit" in sample else []
-    )
+    ind_info["capture_kits"] = [sample.get("capture_kit")] if "capture_kit" in sample else []
 
     # Cancer specific values
     ind_info["tumor_type"] = sample.get("tumor_type")