+++ scout/parse/variant/callers.py
@@ -41,13 +41,11 @@ def parse_callers(variant, category="snv"):
             called_by = info.split("|")[0]
             callers[called_by] = "Pass"
 
-    # Assume GATK was used for calling as a default, and report according to vcf FILTER status.
-    if not raw_info or other_info:
-        filter_status = variant.FILTER
+    if category == "snv" and not raw_info or other_info:
         # cyvcf2 FILTER is None if VCF file column FILTER is "PASS"
-        if filter_status is None:
-            callers["gatk"] = "Pass"
-        else:
-            callers["gatk"] = "Filtered - {}".format(filter_status.replace(";", " - "))
+        filter_status = "Pass"
+        if variant.FILTER is not None:
+            filter_status = "Filtered - {}".format(filter_status.replace(";", " - "))
+        callers["gatk"] = filter_status
 
     return callers