+++ scout/parse/variant/variant.py
@@ -152,9 +152,7 @@ def parse_variant(
 
     ################# Add compound information #################
     compounds = parse_compounds(
-        compound_info=variant.INFO.get("Compounds"),
-        case_id=genmod_key,
-        variant_type=variant_type,
+        compound_info=variant.INFO.get("Compounds"), case_id=genmod_key, variant_type=variant_type,
     )
     if compounds:
         parsed_variant["compounds"] = compounds
@@ -236,9 +234,7 @@ def parse_variant(
     parsed_transcripts = []
     dbsnp_ids = set()
     cosmic_ids = set()
-    for parsed_transcript in parse_transcripts(
-        raw_transcripts, parsed_variant["alternative"]
-    ):
+    for parsed_transcript in parse_transcripts(raw_transcripts, parsed_variant["alternative"]):
         parsed_transcripts.append(parsed_transcript)
         for dbsnp in parsed_transcript.get("dbsnp", []):
             dbsnp_ids.add(dbsnp)