+++ scout/server/app.py
@@ -74,12 +74,8 @@ def create_app(config_file=None, config=None):
     def check_user():
         if not app.config.get("LOGIN_DISABLED") and request.endpoint:
             # check if the endpoint requires authentication
-            static_endpoint = (
-                "static" in request.endpoint or "report" in request.endpoint
-            )
-            public_endpoint = getattr(
-                app.view_functions[request.endpoint], "is_public", False
-            )
+            static_endpoint = "static" in request.endpoint or "report" in request.endpoint
+            public_endpoint = getattr(app.view_functions[request.endpoint], "is_public", False)
             relevant_endpoint = not (static_endpoint or public_endpoint)
             # if endpoint requires auth, check if user is authenticated
             if relevant_endpoint and not current_user.is_authenticated:
@@ -184,8 +180,7 @@ def configure_email_logging(app):
     mail_handler.setLevel(logging.ERROR)
     mail_handler.setFormatter(
         logging.Formatter(
-            "%(asctime)s - %(name)s - %(levelname)s: %(message)s "
-            "[in %(pathname)s:%(lineno)d]"
+            "%(asctime)s - %(name)s - %(levelname)s: %(message)s " "[in %(pathname)s:%(lineno)d]"
         )
     )
     app.logger.addHandler(mail_handler)