+++ scout/server/blueprints/variant/controllers.py
@@ -109,9 +109,7 @@ def variant(
         genome_build = "37"
 
     panels = default_panels(store, case_obj)
-    variant_obj = add_gene_info(
-        store, variant_obj, gene_panels=panels, genome_build=genome_build
-    )
+    variant_obj = add_gene_info(store, variant_obj, gene_panels=panels, genome_build=genome_build)
     # Add information about bam files and create a region vcf
     if add_case:
         variant_case(store, case_obj, variant_obj)
@@ -162,9 +160,7 @@ def variant(
         # This is to convert a summary of frequencies to a string
         variant_obj["frequency"] = frequency(variant_obj)
     # Format clinvar information
-    variant_obj["clinsig_human"] = (
-        clinsig_human(variant_obj) if variant_obj.get("clnsig") else None
-    )
+    variant_obj["clinsig_human"] = clinsig_human(variant_obj) if variant_obj.get("clnsig") else None
 
     # Add display information about callers
     variant_obj["callers"] = callers(variant_obj, category=variant_category)
@@ -173,13 +169,9 @@ def variant(
     is_affected(variant_obj, case_obj)
 
     if variant_obj.get("genetic_models"):
-        variant_models = set(
-            model.split("_", 1)[0] for model in variant_obj["genetic_models"]
-        )
+        variant_models = set(model.split("_", 1)[0] for model in variant_obj["genetic_models"])
         all_models = variant_obj.get("all_models", set())
-        variant_obj["is_matching_inheritance"] = set.intersection(
-            variant_models, all_models
-        )
+        variant_obj["is_matching_inheritance"] = set.intersection(variant_models, all_models)
 
     # Prepare classification information for visualisation
     classification = variant_obj.get("acmg_classification")
@@ -274,9 +266,7 @@ def observations(store, loqusdb, case_obj, variant_obj):
         obs_data["total"] = loqusdb.case_count()
         return obs_data
 
-    user_institutes_ids = set(
-        [inst["_id"] for inst in user_institutes(store, current_user)]
-    )
+    user_institutes_ids = set([inst["_id"] for inst in user_institutes(store, current_user)])
 
     obs_data["cases"] = []
     institute_id = variant_obj["institute"]
@@ -349,10 +339,7 @@ def variant_acmg_post(store, institute_id, case_name, variant_id, user_email, cr
     variant_obj = store.variant(variant_id)
     user_obj = store.user(user_email)
     variant_link = url_for(
-        "variant.variant",
-        institute_id=institute_id,
-        case_name=case_name,
-        variant_id=variant_id,
+        "variant.variant", institute_id=institute_id, case_name=case_name, variant_id=variant_id,
     )
     classification = store.submit_evaluation(
         institute_obj=institute_obj,
@@ -381,14 +368,13 @@ def clinvar_export(store, institute_id, case_name, variant_id):
     """
     institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
     pinned = [
-        store.variant(variant_id) or variant_id
-        for variant_id in case_obj.get("suspects", [])
+        store.variant(variant_id) or variant_id for variant_id in case_obj.get("suspects", [])
     ]
     variant_obj = store.variant(variant_id)
 
     # gather missing transcript info from entrez (refseq id version)
     for pinned_var in pinned:
-        for gene in pinned_var.get("genes"):
+        for gene in pinned_var.get("genes", []):
             for transcript in gene.get("transcripts"):
                 refseq_id = transcript.get("refseq_id")
                 if not refseq_id: