+++ tests/commands/export/test_export_mt_report_cmd.py
@@ -16,9 +16,7 @@ def test_export_mt_report(mock_app, case_obj):
     assert 'Missing option "--case_id"' in result.output
 
     # Test the CLI providing case_id
-    result = runner.invoke(
-        cli, ["export", "mt_report", "--case_id", case_obj["_id"], "--test"]
-    )
+    result = runner.invoke(cli, ["export", "mt_report", "--case_id", case_obj["_id"], "--test"])
     # there are no variants in database so you should get an error message
     assert "There are no MT variants associated to case" in result.output
 
@@ -29,9 +27,7 @@ def test_export_mt_report(mock_app, case_obj):
     assert sum(1 for i in store.variant_collection.find({"chromosome": "MT"})) > 0
 
     # Test the CLI providing case_id
-    result = runner.invoke(
-        cli, ["export", "mt_report", "--case_id", case_obj["_id"], "--test"]
-    )
+    result = runner.invoke(cli, ["export", "mt_report", "--case_id", case_obj["_id"], "--test"])
     assert result.exit_code == 0
     # This time it should return a message saying thate Excel files can be written to outfolder
     assert "INFO Number of excel files that can be written to folder" in result.output