+++ tests/server/blueprints/dashboard/test_dashboard_controllers.py
@@ -54,9 +54,7 @@ def test_with_slice_query(real_adapter, case_obj):
     institute_id = case_obj["owner"]
 
     slice_query = case_display_id
-    data = get_dashboard_info(
-        adapter, institute_id=institute_id, slice_query=slice_query
-    )
+    data = get_dashboard_info(adapter, institute_id=institute_id, slice_query=slice_query)
 
     ## THEN assert there is one case in the data
     for group in data["cases"]:
@@ -81,9 +79,7 @@ def test_with_hpo_query(real_adapter, case_obj):
     institute_id = case_obj["owner"]
     slice_query = phenotype["phenotype_id"]
 
-    data = get_dashboard_info(
-        adapter, institute_id=institute_id, slice_query=slice_query
-    )
+    data = get_dashboard_info(adapter, institute_id=institute_id, slice_query=slice_query)
     ## THEN assert there is one case in the data
     for group in data["cases"]:
         if group["status"] == "all":
@@ -107,9 +103,7 @@ def test_with_phenotype_group_query(real_adapter, case_obj):
     institute_id = case_obj["owner"]
     slice_query = "PG:0000001"
 
-    data = get_dashboard_info(
-        adapter, institute_id=institute_id, slice_query=slice_query
-    )
+    data = get_dashboard_info(adapter, institute_id=institute_id, slice_query=slice_query)
     ## THEN assert there is one case in the data
     for group in data["cases"]:
         if group["status"] == "all":