+++ CHANGELOG.md
@@ -5,9 +5,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## [x.x.x]
-### Added
-
-## [4.16.2]
 ### Ádded
 - Added if statements for UCSC-link from Structural variant view, for hg19/hg38
 
@@ -15,6 +12,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Bug in clinVar form when variant has no gene
 - Bug when sharing cases with the same institute twice
 - Page crashing when removing causative variant tag
+- Do not default to GATK caller when no caller info is provided for cancer SNVs
 
 ### Changed